フォーム送信時の処理をAuthFormコンポーネントに統合、FormTypeによって出し分けるように変更
